Split out ObservableArray
